lib: propose new architecture
The idea of new approach is to create two separate tasks for
handling terminal events and interrupts. The terminal manager task
processes all incoming packages and call corresponding function.
All handlers should be defined in 'command_handlers.c/h' files.
Meanwhile, interrupt manager catches all interrupts and call
appropriate handler for each REGISTERED interrupt.

So far, architecture does not seem to be an entirely platform independent
but soon it will be refined in a way that it will be.

Currently, terminal does not function due to uninitialized neccesary
peripherals. Fixes are on the way.

Signed-off-by: Edgar Kaziakhmedov <edos.edgar@gmail.com>